+++ lib/dqt/api/v1/qualified_teaching_status.rb
@@ -14,6 +14,8 @@ module Dqt
 
           response = client.get(path: "/api/qualified-teachers/qualified-teaching-status", params: mapped_params)
 
+          return [] if response[:data] == nil
+
           # API returns multiple items but we only ever use the first one. Decided to create a consistent interface here for automated checks rather than spend time creating an abstract interface.
           first_item = response[:data].first